refactor: extract native PDF conversion backend
